packages/server/python.org/youtube-dlp-whisper/pyproject.toml
[project]
name = "youtube-dlp-whisper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.14.2"
dependencies = [
    "fastapi==0.124.0",
    "openai-whisper",
    "pydantic==2.12.5",
    "requests==2.32.5",
    "structlog==25.5.0",
    "yt-dlp==2025.12.8",
]
